refactor(mail): reorganize HTML escaping utilities
- Moved the escapeHtml function to a new file, html.ts, and added an unescapeHtml function to complement it.
- Updated references to the new location of escapeHtml in escapeTemplateArguments.
- Removed outdated tests for escapeHtml and added new tests for both escapeHtml and unescapeHtml to ensure functionality.

This change improves the organization of HTML-related utilities and enhances test coverage.
